Cleaner Code
Former-commit-id: 6caea7b9a518f2b92981685401475d9c2c1b5aa3
Former-commit-id: 85864f694455da968dc758ffc17fd639b2262735

+++ srcs/input/move.c
@@ -41,7 +41,7 @@ void		sprint_release(t_doom *doom)
 	doom->player.speed = 10;
 }
 
-void		move(t_doom *doom, int x, int y)
+void		move(t_doom *doom, double x, double y)
 {
 	t_fvct2 d;
 	t_fvct2 npos;
